Add plugin to minify HTML when in production
Part of the Gulp task that was ran during deploy was minifying a bunch
of things. Jekyll assets takes care of the assets, but nothing in Ruby
was doing the HTML. I added a plugin that minifies the HTML when running
in production.

I also noticed that the Gemfile was using Windows line endings instead
of Linux endings, so I converted that file to use normal line endings.