Fix deploy
Updated the travis and s3_website configuration to work with the new way
the site is being built.

Added some content to README on how to install, build, and test the
site.

Added some TODOs to the README as well. These are mostly things that
were cool to have and worked when we had Gulp, but no longer do since
moving to a full Ruby setup.