Initial commit of new theme
The old theme had some issues on mobile and with long posts, having all content on the page got a little distracting.

For the new theme, there were a few things I wanted from it. First, I wanted to use PostCSS. This seems to be the direction front ends are going. Second, I didn't want to use Bootstrap because I always use Bootstrap.

Skeleton looked pretty cool for a framework, but it hasn't been updated in a while and isn't on NPM. Milligram was my second choice, but it had a very limited browser support (only current browsers). I landed on PureCSS. It's a small framework that doesn't have a lot, but is also compatible with other frameworks. I used their blog template, which is pretty close to the look I wanted for the site.

I wanted to use Webpack, but that seems to be over my head so I ended up using Gulp, which is still pretty awesome. I removed Jekyll assets because Gulp gives me everything I need. I also removed the compress layout as this is also handled with Gulp and I was having some issues with the compress layout with the new theme.