Mark old posts via JavaScript
Because this site only gets updated when I make a change, the old post message may not appear when it's expected to. To solve this, I moved the message to render via JavaScript, so that if the site doesn't get deployed, the old post messages will still appear when expected.